Refactoring the layouts
