--- packages/strapi-parts/src/GridLayout/index.js
@@ -1,1 +1,0 @@
-export * from './GridLayout';
