--- packages/strapi-parts/src/HeaderLayout/index.js
@@ -1,1 +1,0 @@
-export * from './HeaderLayout';
