+++ packages/strapi-parts/src/Layout/GridLayout.stories.mdx
@@ -1,0 +1,31 @@
+<!--- GridLayout.stories.mdx --->
+
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
+import { GridLayout } from './GridLayout';
+import { Box } from '../Box';
+
+<Meta title="Design System/Layouts/GridLayout" component={GridLayout} />
+
+# GridLayout
+
+This is the doc of the `GridLayout` component
+
+## GridLayout
+
+Description...
+
+<Canvas>
+  <Story name="base">
+    <Box padding={8} background="neutral100">
+      <GridLayout>
+        {Array(12)
+          .fill(null)
+          .map((_, idx) => (
+            <Box padding={4} hasRadius background="neutral0" padding={4} key={`box-${idx}`} shadow="tableShadow">
+              hello world
+            </Box>
+          ))}
+      </GridLayout>
+    </Box>
+  </Story>
+</Canvas>
