+++ packages/strapi-parts/src/Layout/Layout.stories.mdx
@@ -1,0 +1,213 @@
+<!--- Layout.stories.mdx --->
+
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
+import { Applications, ApiTokens, AlertWarningIcon, AlertInfoIcon, AddIcon, EditIcon, BackIcon } from '@strapi/icons';
+import DeleteIcon from '@strapi/icons/DeleteIcon';
+import { MemoryRouter } from 'react-router-dom';
+import { SubNav, SubNavHeader, SubNavSection, SubNavSections, SubNavLink, SubNavLinkSection } from '../SubNav';
+import { Link } from '../Link';
+import { Layout } from './Layout';
+import { Box } from '../Box';
+import { Button } from '../Button';
+import { HeaderLayout } from './HeaderLayout';
+import { ActionLayout } from './ActionLayout';
+import { ContentLayout } from './ContentLayout';
+import { Row } from '../Row';
+import { Table, Thead, Tbody, Tr, Td, Th, TFooter } from '../Table';
+import { TableLabel, Text } from '../Text';
+import { BaseCheckbox } from '../BaseCheckbox';
+import { IconButton } from '../IconButton';
+
+<Meta title="Design System/Layouts/Layout" component={Layout} />
+
+# Layout
+
+This is the doc of the `Layout` component
+
+## Layout
+
+Description...
+
+<Canvas>
+  <Story name="sidenav">
+    {() => {
+      const links = [
+        {
+          id: 1,
+          label: 'Addresses',
+          icon: <AlertWarningIcon />,
+          to: '/address',
+        },
+        {
+          id: 2,
+          label: 'Categories',
+          to: '/category',
+        },
+        {
+          id: 3,
+          label: 'Cities',
+          icon: <Applications />,
+          to: '/city',
+          active: true,
+        },
+        {
+          id: 4,
+          label: 'Countries',
+          to: '/country',
+        },
+      ];
+      const ROW_COUNT = 6;
+      const COL_COUNT = 10;
+      const entry = {
+        cover: 'https://avatars.githubusercontent.com/u/3874873?v=4',
+        description: 'Chez Léon is a human sized Parisian',
+        category: 'French cuisine',
+        contact: 'Leon Lafrite',
+      };
+      const entries = [];
+      for (let i = 0; i < 5; i++) {
+        entries.push({ ...entry, id: i });
+      }
+      return (
+        <Box background="neutral100">
+          <Layout
+            sideNav={
+              <SubNav ariaLabel="Builder sub nav">
+                <SubNavHeader
+                  searchable
+                  value={''}
+                  onClear={() => {}}
+                  onChange={(e) => {}}
+                  label="Builder"
+                  searchLabel="Search..."
+                />
+                <SubNavSections>
+                  <SubNavSection label="Collection Type" collapsable badgeLabel={links.length.toString()}>
+                    {links.map((link) => (
+                      <SubNavLink to={link.to} active={link.active} key={link.id}>
+                        {link.label}
+                      </SubNavLink>
+                    ))}
+                  </SubNavSection>
+                  <SubNavSection label="Single Type" collapsable badgeLabel={links.length.toString()}>
+                    <SubNavLinkSection label="Default">
+                      {links.map((link) => (
+                        <SubNavLink to={link.to} key={link.id}>
+                          {link.label}
+                        </SubNavLink>
+                      ))}
+                    </SubNavLinkSection>
+                  </SubNavSection>
+                </SubNavSections>
+              </SubNav>
+            }
+          >
+            <>
+              <HeaderLayout
+                primaryAction={<Button startIcon={<AddIcon />}>Add an entry</Button>}
+                secondaryAction={
+                  <Button variant="tertiary" startIcon={<EditIcon />}>
+                    Edit
+                  </Button>
+                }
+                title="Other CT"
+                subtitle="36 entries found"
+                as="h2"
+              />
+              <ActionLayout
+                startActions={
+                  <>
+                    <Button variant="tertiary">Search</Button>
+                    <Button variant="tertiary">Filter</Button>
+                  </>
+                }
+                endActions={
+                  <>
+                    <Button variant="tertiary">Settings</Button>
+                  </>
+                }
+              />
+              <ContentLayout>
+                <Table
+                  colCount={COL_COUNT}
+                  rowCount={ROW_COUNT}
+                  footer={<TFooter icon={<AddIcon />}>Add another field to this collection type</TFooter>}
+                >
+                  <Thead>
+                    <Tr>
+                      <Th>
+                        <BaseCheckbox aria-label="Select all entries" />
+                      </Th>
+                      <Th>
+                        <TableLabel>ID</TableLabel>
+                      </Th>
+                      <Th>
+                        <TableLabel>Cover</TableLabel>
+                      </Th>
+                      <Th>
+                        <TableLabel>Description</TableLabel>
+                      </Th>
+                      <Th>
+                        <TableLabel>Categories</TableLabel>
+                      </Th>
+                      <Th>
+                        <TableLabel>Contact</TableLabel>
+                      </Th>
+                      <Th>More</Th>
+                      <Th>More</Th>
+                      <Th>More</Th>
+                      <Th>
+                        <VisuallyHidden>Actions</VisuallyHidden>
+                      </Th>
+                    </Tr>
+                  </Thead>
+                  <Tbody>
+                    {entries.map((entry) => (
+                      <Tr key={entry.id}>
+                        <Td>
+                          <BaseCheckbox aria-label={`Select ${entry.contact}`} />
+                        </Td>
+                        <Td>
+                          <Text textColor="neutral800">{entry.id}</Text>
+                        </Td>
+                        <Td>
+                          <Avatar src={entry.cover} alt={entry.contact} />
+                        </Td>
+                        <Td>
+                          <Text textColor="neutral800">{entry.description}</Text>
+                        </Td>
+                        <Td>
+                          <Text textColor="neutral800">{entry.category}</Text>
+                        </Td>
+                        <Td>
+                          <Text textColor="neutral800">{entry.contact}</Text>
+                        </Td>
+                        <Td>
+                          <Text textColor="neutral800">{entry.description}</Text>
+                        </Td>
+                        <Td>
+                          <Text textColor="neutral800">{entry.description}</Text>
+                        </Td>
+                        <Td>
+                          <Text textColor="neutral800">{entry.description}</Text>
+                        </Td>
+                        <Td>
+                          <Row>
+                            <IconButton onClick={() => console.log('edit')} label="Edit" icon={<EditIcon />} />
+                            <Box paddingLeft={1}>
+                              <IconButton onClick={() => console.log('edit')} label="Delete" icon={<DeleteIcon />} />
+                            </Box>
+                          </Row>
+                        </Td>
+                      </Tr>
+                    ))}
+                  </Tbody>
+                </Table>
+              </ContentLayout>
+            </>
+          </Layout>
+        </Box>
+      );
+    }}
+  </Story>
+</Canvas>
