+++ packages/strapi-parts/src/Layout/__tests__/TwoColsLayout.e2e.js
