+++ packages/strapi-parts/src/Layout/index.js
@@ -1,0 +1,5 @@
+export * from './Layout';
+export * from './ActionLayout';
+export * from './ContentLayout';
+export * from './HeaderLayout';
+export * from './TwoColsLayout';
