--- packages/strapi-parts/src/OneBlockLayout/index.js
@@ -1,1 +1,0 @@
-export * from './OneBlockLayout';
