--- packages/strapi-parts/src/TwoColsLayout/index.js
@@ -1,1 +1,0 @@
-export * from './TwoColsLayout';
