+++ packages/strapi-parts/src/index.js
@@ -18,16 +18,14 @@ export * from './Field';
 export * from './FocusTrap';
 export * from './GlobalStyle';
 export * from './Grid';
-export * from './GridLayout';
-export * from './HeaderLayout';
 export * from './IconButton';
 export * from './KeyboardNavigable';
+export * from './Layout';
 export * from './Link';
 export * from './Loader';
 export * from './Main';
 export * from './MainNav';
 export * from './ModalLayout';
-export * from './OneBlockLayout';
 export * from './Pagination';
 export * from './Popover';
 export * from './Portal';
@@ -54,6 +52,5 @@ export * from './ThemeProvider';
 export * from './TimePicker';
 export * from './ToggleCheckbox';
 export * from './Tooltip';
-export * from './TwoColsLayout';
 export * from './VisuallyHidden';
 export * from './themes';
